+++ examples/substrate-to-evm-fungible-transfer/src/transfer.ts
@@ -44,7 +44,6 @@ const substrateTransfer = async (): Promise<void> => {
     resource: RESOURCE_ID_SYGMA_USD,
     amount: BigInt("1"),
     destinationAddress: recipient,
-    sourceAddress: account.address,
   };
 
   const transfer = await createSubstrateFungibleAssetTransfer(transferParams);